requirements.txt
openai>=1.0
anthropic>=0.20
google-generativeai>=0.5
httpx>=0.27
sentence-transformers>=2.2
hdbscan>=0.8.1
scikit-learn>=1.3
pydantic>=2.0
python-dotenv>=1.0
numpy>=1.24
fastapi>=0.100
uvicorn>=0.20
python-multipart>=0.0.6

